backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
requests
pydantic-settings
sqlalchemy
python-dotenv
